requirements.txt
kafka-python
elasticsearch
flask
python-dotenv
requests
uuid